[Utils] split into math and system utilities
